refactor(step4): 패키지 구조 변경
dto 따로 패키징
test 위치와 대상 위치를 같이하기
