+++ src/main/java/lotto/dto/ManualLottoOrderSheet.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import java.util.ArrayList;
 import java.util.List;
