+++ src/test/java/lotto/domain/LottoInspectorTest.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import lotto.domain.*;
+import lotto.domain.InsightResults;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
