+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.domain.Money;
 import org.junit.jupiter.api.DisplayName;
