+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.domain.WinningLotto;
 import org.junit.jupiter.api.DisplayName;
